# Parallel v2: Change parallelization approach (`parallel` to `n_threads`)