--- a/src/force.jl
+++ b/src/force.jl
@@ -221,54 +221,65 @@ Custom general interaction types should implement this function.
 function forces(s::System{D, false}, neighbors=nothing; n_threads::Integer=Threads.nthreads()) where D
     n_atoms = length(s)
 
-    # allocate only `missing` for unused threads instead of a large array
-    T = typeof(ustrip_vec.(s.coords))
-    fs_threads = Union{Missing, T}[missing for i in 1:Threads.nthreads()]
-    thread_used = falses(Threads.nthreads())
-
-    for inter in values(s.pairwise_inters)
-        if inter.nl_only
-            if isnothing(neighbors)
-                error("An interaction uses the neighbor list but neighbors is nothing")
-            end
-            @floop ThreadedEx(basesize = neighbors.n ÷ n_threads) for ni in 1:neighbors.n
-                FLoops.@init id = Threads.threadid()
-                if !thread_used[id]
-                    fs_threads[id] = ustrip_vec.(zero(s.coords))
-                    thread_used[id] = true
+    if n_threads > 1
+        fs_chunks = [ustrip_vec.(zero(s.coords)) for i=1:n_threads]
+        for inter in values(s.pairwise_inters)
+            if inter.nl_only
+                if isnothing(neighbors)
+                    error("An interaction uses the neighbor list but neighbors is nothing")
                 end
-                i, j, w = neighbors.list[ni]
-                Molly.force!(fs_threads[id], inter, s, i, j, s.force_units, w)
-            end
-        else
-            @floop ThreadedEx(basesize = n_atoms ÷ n_threads) for i in 1:n_atoms
-                FLoops.@init id = Threads.threadid()
-                if !thread_used[id]
-                    fs_threads[id] = ustrip_vec.(zero(s.coords))
-                    thread_used[id] = true
+                basesize = max(1, Int(ceil(neighbors.n / n_threads)))
+                chunks = [i:min(i + basesize - 1, neighbors.n) for i in 1:basesize:neighbors.n]
+                # @threads need collect to be used
+                Threads.@threads for (id, rng) in collect(enumerate(chunks))
+                    for ni in rng
+                        i, j, w = neighbors.list[ni]
+                        Molly.force!(fs_chunks[id], inter, s, i, j, s.force_units, w)
+                    end
                 end
-                for j in 1:(i - 1)
-                    Molly.force!(fs_threads[id], inter, s, i, j, s.force_units)
+            else
+                basesize = max(1, Int(ceil(n_atoms / n_threads)))
+                chunks = [i:min(i + basesize - 1, n_atoms) for i in 1:basesize:n_atoms]
+                Threads.@threads for (id, rng) in collect(enumerate(chunks))
+                    for i in rng
+                        for j in 1:(i - 1)
+                            Molly.force!(fs_chunks[id], inter, s, i, j, s.force_units)
+                        end
+                    end
                 end
             end
         end
-    end
-
-    fs = ustrip_vec.(zero(s.coords))
-    for id in 1:Threads.nthreads()
-        if thread_used[id]
-            fs .+= fs_threads[id]
+    
+        fs = sum(fs_chunks)
+    else
+        fs = ustrip_vec.(zero(s.coords))
+        for inter in values(s.pairwise_inters)
+            if inter.nl_only
+                if isnothing(neighbors)
+                    error("An interaction uses the neighbor list but neighbors is nothing")
+                end
+                for ni in 1:neighbors.n
+                    i, j, w = neighbors.list[ni]
+                    Molly.force!(fs, inter, s, i, j, s.force_units, w)
+                end
+            else
+                for i in 1:n_atoms
+                    for j in 1:(i - 1)
+                        Molly.force!(fs, inter, s, i, j, s.force_units)
+                    end
+                end
+            end
         end
     end
 
     for inter_list in values(s.specific_inter_lists)
         sparse_vec = specific_force(inter_list, s.coords, s.boundary, s.force_units, n_atoms)
-        fs .+= ustrip_vec.(Array(sparse_vec))
+        fs += ustrip_vec.(Array(sparse_vec))
     end
 
     for inter in values(s.general_inters)
         # Force type not checked
-        fs .+= ustrip_vec.(forces(inter, s, neighbors))
+        fs += ustrip_vec.(forces(inter, s, neighbors))
     end
 
     return fs * s.force_units