The following changes are made in the PR:
1. The kwarg `parallel` is replace by `n_threads` throughout the library which can be used to specify the number of threads.
2. Use of `@floop` instead of `@threads` for the use of `n_threads` .

Changes yet to be done:
- [ ] Use reductions in `forces`  (still uses `Threads.@threads` for passing tests)
- [x] Use reductions in `find_neighbors`
- [ ] Possible simplification in `log_property!` method for `TimeCorrelationLogger`