@jgreener64 2022-07-04
Looks great. Yes that change to `find_neighbors` is fine. I would prioritise the `forces` reduction over `find_neighbors` and `TimeCorrelationLogger` since that sees the most use.

It would be useful to see if the benchmark results change with these changes. This would involve running `benchmark/benchmarks.jl` (instructions at the top of the file) and `benchmark/protein.jl` (which is just a script) on a machine with multiple cores with and without this PR. As long as performance doesn't drop then it's fine. Don't worry about the GPU benchmarks.