@JaydevSR 2022-07-08
I tried using reductions for `forces` function, but It was turning out to be too complicated. So I ended up with this in this routine in the end:
```julia
function forces(s::System{D, false}, neighbors=nothing; n_threads::Integer=Threads.nthreads()) where D
    n_atoms = length(s)

    # allocate only `missing` for unused threads instead of a large array
    T = typeof(ustrip_vec.(s.coords))
    fs_threads = Union{Missing, T}[missing for i in 1:Threads.nthreads()]
    thread_used = falses(Threads.nthreads())

    for inter in values(s.pairwise_inters)
        if inter.nl_only
            if isnothing(neighbors)
                error("An interaction uses the neighbor list but neighbors is nothing")
            end
            @floop ThreadedEx(basesize = neighbors.n ÷ n_threads) for ni in 1:neighbors.n
                FLoops.@init id = Threads.threadid()
                if !thread_used[id]
                    fs_threads[id] = ustrip_vec.(zero(s.coords))
                    thread_used[id] = true
                end
                i, j, w = neighbors.list[ni]
                Molly.force!(fs_threads[id], inter, s, i, j, s.force_units, w)
            end
        else
            @floop ThreadedEx(basesize = n_atoms ÷ n_threads) for i in 1:n_atoms
                FLoops.@init id = Threads.threadid()
                if !thread_used[id]
                    fs_threads[id] = ustrip_vec.(zero(s.coords))
                    thread_used[id] = true
                end
                for j in 1:(i - 1)
                    Molly.force!(fs_threads[id], inter, s, i, j, s.force_units)
                end
            end
        end
    end

    fs = ustrip_vec.(zero(s.coords))
    for id in 1:Threads.nthreads()
        if thread_used[id]
            fs .+= fs_threads[id]
        end
    end

    for inter_list in values(s.specific_inter_lists)
        sparse_vec = specific_force(inter_list, s.coords, s.boundary, s.force_units, n_atoms)
        fs .+= ustrip_vec.(Array(sparse_vec))
    end

    for inter in values(s.general_inters)
        # Force type not checked
        fs .+= ustrip_vec.(forces(inter, s, neighbors))
    end

    return fs * s.force_units
end
```

I was thinking of using `fs = sum(skipmissing(fs_threads)` before but it was not working as in some cases `sum` requires the zero of the type it is summing over which is not defined on our case. Due to this I had the add `threads_used` vector. This saves on allocations and time for smaller number of threads used.

The only other problem that I am seeing right now is that, the test https://github.com/JuliaMolSim/Molly.jl/blob/3b5d2b1f4808fab8570f2c9b307dd11dc63c3c33/test/simulation.jl#L365 is failing. Which after some discussion on julia slack, I think is because `@floops` dynamically schedules the tasks so the order of tasks on threads is not same for each run. This leads to differences of the order of `eps()` which is equal to `2.220446049250313e-16` for me as floating point arithmatic is not associative (the value that `sum(accel_diff)` returns is `Quantity{Float64, 𝐋 𝐍^-1 𝐓^-2, Unitful.FreeUnits{(kJ, nm^-1, mol^-1, u^-1), 𝐋 𝐍^-1 𝐓^-2, nothing}}[-3.1712913545201005e-16 kJ nm^-1 mol^-1 u^-1, 2.207435623180487e-16 kJ nm^-1 mol^-1 u^-1, -2.2724877535296173e-16 kJ nm^-1 mol^-1 u^-1]` where all elements are of order `eps()`). Maybe the test can be rewritten to check values elementwise to be approximate if this is not a very big issue from simulation perspective. Something like this:

```julia
a1 = accelerations(s, neighbors)
a2 = accelerations(s_nounits, neighbors_nounits)u"kJ * mol^-1 * nm^-1 * u^-1"
@assert all(all(a1[i] .≈ a2[i]) for i in eachindex(a1)) == true
```

@jgreener64 If I can go ahead with this then I will commit these changes and report the benchmark.