@jgreener64 2022-07-08
Yes give it a go, I'd be interested in seeing the benchmarks. Changing that test is fine.